// src/app/model/users.ts
export class Users {
    id?: number;
    name: string;
    surname: string;
    number: string; 
    photo: string;
    image: string;
    isFavorite: boolean;

constructor() {
    this.name = '';
    this.surname = '';
    this.number = '';
    this.photo = '';
    this.image = '';
    this.isFavorite = false;

}
}